Renamed HSPIDevice folders to hspi_
