+++ BREAKING_CHANGES.md
@@ -2,3 +2,5 @@ v1.1.4:
 
 * `USBDevice` folder renamed to `usb`
 * `SerDesDevice` folder renamed to `serdes`
+* `HSPIDeviceScheduled` folder renamed to `hspi_scheduled`
+* `HSPIDevice` folder renamed to `hspi`
