+++ src/CMakeLists.txt
@@ -1,9 +1,9 @@
 add_library(wch-ch56x-lib INTERFACE)
 
-# With HSPIDeviceScheduled
+# With hspi_scheduled
 target_sources(wch-ch56x-lib INTERFACE
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/utils/critical_section.c
-    ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/HSPIDevice/hspi.c
+    ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/hspi/hspi.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/interrupt_queue/interrupt_queue.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/logging/log_printf.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/logging/log_serdes.c
@@ -23,13 +23,13 @@ target_include_directories(wch-ch56x-lib INTERFACE ${CMAKE_CURRENT_LIST_DIR})
 
 target_link_libraries(wch-ch56x-lib INTERFACE wch-ch56x-bsp lwrb nanoprintf)
 
-# Without HSPIDeviceScheduled
+# Without hspi_scheduled
 
 add_library(wch-ch56x-lib-scheduled INTERFACE)
 
 target_sources(wch-ch56x-lib-scheduled INTERFACE
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/utils/critical_section.c
-    ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/HSPIDeviceScheduled/hspi_scheduled.c
+    ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/hspi_scheduled/hspi_scheduled.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/interrupt_queue/interrupt_queue.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/logging/log_printf.c
     ${CMAKE_CURRENT_LIST_DIR}/wch-ch56x-lib/logging/log_serdes.c
