+++ tests/test_firmware_hspi/User/main.c
@@ -26,7 +26,7 @@ limitations under the License.
 #pragma GCC diagnostic pop
 #pragma GCC diagnostic pop
 
-#include "wch-ch56x-lib/HSPIDevice/hspi.h"
+#include "wch-ch56x-lib/hspi/hspi.h"
 #include "wch-ch56x-lib/logging/logging.h"
 #include "wch-ch56x-lib/usb/usb20.h"
 #include "wch-ch56x-lib/usb/usb30.h"
